♻️ Refactor :  footer 폴더 이동 & PATH 상수 사용
