+++ src/shared/constants/path.ts
@@ -1,5 +1,7 @@
 const PATH = {
   home: '/',
+  search: '/search',
+  allProductList: '/products',
   wishProductList: '/wishlist/list/products',
   wishList: '/wishlist/list',
   wishLogin: '/wishlist/login',
